Move set visitor to module
